requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
keras==2.13.1
Pillow==10.0.0
numpy==1.24.3
tensorflow==2.13.0
Jinja2==3.1.2
python-multipart==0.0.6
opencv-python==4.8.1.78
scipy==1.11.2
sqlalchemy==1.4.25